+++ src/vue/mixins/verification-form.mixin.js
@@ -19,11 +19,19 @@ export default {
       kycRequestId: vuexTypes.kycRequestId,
       accountId: vuexTypes.accountId,
     }),
+
+    isUpdatableKycRequest () {
+      return (
+        this.kycState === REQUEST_STATES_STR.rejected ||
+        this.kycState === REQUEST_STATES_STR.pending
+      )
+    },
   },
   methods: {
     ...mapActions({
       loadKyc: vuexTypes.LOAD_KYC,
     }),
+
     async createKycBlob (blobType) {
       const { data } = await Sdk.api.blobs.create(
         blobType,
@@ -31,9 +39,10 @@ export default {
       )
       return data.id
     },
+
     createKycOperation (kycBlobId, accountRole) {
       return base.CreateChangeRoleRequestBuilder.createChangeRoleRequest({
-        requestID: this.kycState === REQUEST_STATES_STR.rejected
+        requestID: this.isUpdatableKycRequest
           ? this.kycRequestId
           : KYC_CREATION_REQUEST_ID,
         destinationAccount: this.accountId,
@@ -42,6 +51,13 @@ export default {
           blob_id: kycBlobId,
         },
       })
+    },
+
+    getKycOperationRequestId () {
+      const isExistingRequest =
+        this.kycState === REQUEST_STATES_STR.rejected ||
+        this.kycState === REQUEST_STATES_STR.pending
+      return isExistingRequest ? this.kycRequestId : KYC_CREATION_REQUEST_ID
     },
   },
 }